fix : 회원수정이 안되는 문제가 생겨서 확인해보니 mem 더미 데이터에 생일이 빠져서 오류가 발생 다시 추가하여 넣었습니다.

+++ OneTouch/src/main/java/com/onetouch/service/MemServiceImpl.java
@@ -104,9 +104,11 @@ public class MemServiceImpl implements MemService {
 			throw new Exception();
 		}
 		if(memVo.getMem_image()!=null&&!memVo.getMem_image().isEmpty()) {
-			System.out.println("		기존 이미지 파일 삭제");
-		File ff =new File(absPath,OriginImageName);
-		ff.delete();
+			if (OriginImageName != null && !OriginImageName.trim().isEmpty()) {
+				System.out.println("		기존 이미지 파일 삭제");
+				File ff =new File(absPath,OriginImageName);
+				ff.delete();
+			}
 		}
 		return res;
 	}
